✨ Add private award management hooks and refactor award module for improved functionality
